requirements.txt
httpx
fastapi
uvicorn 
jinja2
python-multipart
pydantic
sqlalchemy
python-dotenv
python-jose
bcrypt==4.0.1